+++ NEWS.md
@@ -1,3 +1,11 @@
+# Development
+
+## Minor changes
+
+- Add a rate limit checker for the GitHub API calls. Will warn if rate limit is exceeded.
+
+---
+
 # v 0.1.1
 
 ## Minor changes